backend/core/utils/AI.py
import openai


openai.api_key = 'your-api-key'


def parse_llm_output(output):
    # This is a placeholder function. You'll need to implement a parser that can extract the question, answer, and distractors from the model's output.
    # The actual implementation will depend on the format of the model's output.
    pass


class AI:
    def __init__(self, data, model="gpt-4-turbo"):
        self.model = model
        self.data = data
        

    def generate_questions_with_distractors(self):
        response = openai.ChatCompletion.create(
            model="gpt-4-turbo",  # or "gpt-4-turbo-instruct" for more structured outputs
            messages=[
                {"role": "system", "content": "You are a helpful AI that generates quiz questions."},
                {
                    "role": "user",
                    "content": f"""Please create multiple-choice questions based on the following text:
                    Text: \"{self.data}\"
                    
                    For each question, provide:
                    - The question text
                    - The correct answer
                    - Three plausible distractors
                    """
                }
            ],
            max_tokens=500
        )

        questions = []
        if response.choices:
            generated_text = response.choices[0].message['content']
            
            # Parse the response (assuming the model's output is structured in a list format for each question)
            for question_data in parse_llm_output(generated_text):
                questions.append(question_data)
        return questions 