requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
diff-match-patch==20241021
Django==5.1.2
django-admin-kubi==1.1.0
django-admin-tabler==1.0.2
django-grappelli==4.0.1
django-htmx==1.21.0
django-import-export==4.2.0
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-json-widget==2.0.1
django-seed==0.3.1
django-taggit==6.1.0
django-unfold==0.40.0
et_xmlfile==2.0.0
executing==2.1.0
Faker==30.8.2
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
huggingface-hub==0.26.2
icecream==2.1.3
idna==3.10
keras==3.6.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
protobuf==5.28.3
psycopg2==2.9.10
pycparser==2.22
Pygments==2.18.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
six==1.16.0
sqlparse==0.5.1
tablib==3.7.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tokenizers==0.20.3
toposort==1.10
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
XlsxWriter==3.2.0
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
diff-match-patch==20241021
Django==5.1.2
django-admin-kubi==1.1.0
django-admin-tabler==1.0.2
django-grappelli==4.0.1
django-htmx==1.21.0
django-import-export==4.2.0
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-json-widget==2.0.1
django-seed==0.3.1
django-taggit==6.1.0
django-unfold==0.40.0
et_xmlfile==2.0.0
executing==2.1.0
Faker==30.8.2
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
huggingface-hub==0.26.2
icecream==2.1.3
idna==3.10
keras==3.6.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
protobuf==5.28.3
psycopg2==2.9.10
pycparser==2.22
Pygments==2.18.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rules==3.5
safetensors==0.4.5
six==1.16.0
sqlparse==0.5.1
tablib==3.7.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tokenizers==0.20.3
toposort==1.10
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
XlsxWriter==3.2.0
